Refaktoryzacja: Rozdzielenie kodu na osobne moduły
