+++ assets/custom.css
@@ -1,4 +1,4 @@
-/* styl dropdowna w trybie ciemnym */
+ /*styl dropdowna w trybie ciemnym*/
 .dark-dropdown .Select-control {
     background-color: #2a2a2a !important;
     color: white !important;
@@ -22,4 +22,10 @@
 
 .dark-dropdown .Select-placeholder {
     color: #aaa !important;
+}
+
+.dark-dropdown .Select-multi-value-wrapper .Select-value {
+    background-color: #555 !important;
+    color: white !important;
+    border: 1px solid #444 !important;
 }
